nit: remove unneeded float casts

+++ internal/model/modelindexer/indexer.go
@@ -312,11 +312,11 @@ func encodeAny(v interface{}, out *fastjson.Writer) error {
 	case float64:
 		// truncate the value with int conversion to
 		// check for decimals.
-		if float64(v) == float64(int64(v)) {
+		if v == float64(int64(v)) {
 			// we cannot rely on fastjson.Marshal as it would strip
 			// the leading zero and encode the value as an "int".
 			// So we call AppendFloat directly with precision=1.
-			out.RawBytes(strconv.AppendFloat(nil, float64(v), 'f', 1, 64))
+			out.RawBytes(strconv.AppendFloat(nil, v, 'f', 1, 64))
 			return nil
 		}
 		return fastjson.Marshal(out, v)
